pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rra-module"
version = "0.1.0"
description = "Revenant Repo Agent Module - Transform dormant repos into autonomous licensing agents"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "RRA Contributors"}
]
keywords = ["blockchain", "ai", "licensing", "github", "automation", "web3"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
    "gitpython>=3.1.40",
    "requests>=2.31.0",
    "click>=8.1.7",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "web3>=6.11.3",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "httpx>=0.25.2",
    "eth-abi>=4.0.0",
    "eth-utils>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]

# NatLangChain ecosystem integration (optional)
# Install with: pip install rra-module[natlangchain]
natlangchain = [
    # These will be actual packages once the ecosystem repos are published
    # For now, these are placeholder references
    # "natlangchain-common>=0.1.0",
    # "natlangchain-agent-os>=0.1.0",
    # "natlangchain-memory-vault>=0.1.0",
    # "natlangchain-value-ledger>=0.1.0",
    # "natlangchain-mediator-node>=0.1.0",
    # "natlangchain-intent-log>=0.1.0",
    # "natlangchain-synth-mind>=0.1.0",
]

# Full installation with all extras
all = [
    "rra-module[dev,natlangchain]",
]

[project.scripts]
rra = "rra.cli.main:cli"

[project.urls]
Homepage = "https://github.com/kase1111-hash/RRA-Module"
Repository = "https://github.com/kase1111-hash/RRA-Module"
Issues = "https://github.com/kase1111-hash/RRA-Module/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
